add token_owner, api slots
We may want to set the ```user``` slot to someone else (e.g. someone else's repo (or
fork, etC) that we're looking at) such that things Just Work.  It might also
be useful to retain who actually owns the token...  So that's stashed away
independently in ```token_owner``` at construction time.

We also stash a ref to the original instance returned by GitHub::API->new() in
the ```api``` slot, just in case we need it for something new.